reorganized the directories and  removed package folders because classes in other folders were not being imported
