+++ BootCampWeek2_Project/dojo/dojo.py
@@ -1,7 +1,4 @@
 """ Dojo Space Allocation
-
-Welcome to the Dojo Space Allocation system. Run these commands to carry out your desired task
-
 Usage:
     dojo.py create_room <room_type> <room_name>...
     dojo.py add_person <person_name> <FELLOW|STAFF> [wants_accommodation]
@@ -16,10 +13,11 @@ Usage:
         -o,--otherflag  : example of flag #3
 """
 
-import sys
 import cmd
-from docopt import docopt, DocoptExit
-from room.room import Room
+
+from docopt import docopt
+
+from room import Room
 
 """Dojo class which is the main class of the system. It calls and implements most of the
 functionality needed in the system"""
@@ -44,6 +42,27 @@ class Dojo(cmd.Cmd):
         """Adds a person to the system and allocates the person a random room"""
         pass
 
+    def main(self, docopt_args):
+        """ main-entry point for program, expects dict with arguments from docopt() """
+        # User passed the required argument
+        if docopt_args["<required>"]:
+            print("You have used the required argument: " + docopt_args["<required>"])
+
+            # Get flags used
+            if docopt_args["--flag"]:
+                print("   with --flag\n")
+            elif docopt_args["--greatflag"]:
+                print("   with --greatflag\n")
+            elif docopt_args["--otherflag"]:
+                print("   with --otherflag\n")
+            else:
+                print("   with no flags.\n")
+        # User passed 1 or more repeating arguments
+        elif docopt_args["<repeating>"]:
+            print("You have used the repeating args:")
+            print('   ' + '\n   '.join(docopt_args["<repeating>"]) + '\n')
+
     if __name__ == '__main__':
-        arguments = docopt(__doc__, version='Dojo Space Allocation 1.0')
-        print(arguments)
+        args = docopt(__doc__)
+        #arguments = docopt(__doc__, version='Dojo Space Allocation 1.0')
+        main(args)
